refactor: restructure search results page directory
